+++ test/functional/repository/basic-methods/model/User.ts
@@ -1,5 +1,0 @@
-export interface User {
-    id?: number|null;
-    firstName?: string;
-    secondName?: string;
-}